move map packaging function out of map_builder
